Deployments seem OK on Fly
So, while docker doesn't give a shit about non-root users in the
container running processes on ports < 1024, Firecracker does -
and so the deployment was breaking. Oh well. I could probably work
around that in some way, maybe by running a simple healthcheck typa
thing on 80 with nginx and letting the quart app run under a user
but that's overdoing it already - at least it's up and running now.